Rename files to login and home
