+++ client/src/home/home.jsx
@@ -1,0 +1,12 @@
+/* eslint-disable no-underscore-dangle, no-unused-vars */
+import React from 'react';
+
+function Home(props) {
+  return (
+    <div>
+      Landing Page
+    </div>
+  );
+}
+
+export default Home;
